backend/agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "soundsplit",
    "soundgen",
    "google-adk>=1.27.2",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "python-multipart>=0.0.9",
    "motor>=3.3",
    "snowflake-connector-python>=3.0",
]

[project.scripts]
wonder-agent = "agent.server:main"

[tool.uv.sources]
soundsplit = { workspace = true }
soundgen   = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["agent"]
